+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fully clean extension attributes, spans, and direct text nodes in video title elements when the videoId changes to prevent concatenated titles caused by
+
 ## [2.4.24] - 2025-06-22
 
 ### Fixed